dispatcher - memory storage merge WIP
