+++ integration/c/CMakeLists.txt
@@ -16,7 +16,6 @@ set(otterbrix_LIBS
         otterbrix::cursor
         otterbrix::session
         otterbrix::document
-        otterbrix::dispatcher
         otterbrix::wal
         otterbrix::disk
         otterbrix::locks
