+++ integration/c/test/CMakeLists.txt
@@ -20,7 +20,6 @@ target_link_libraries(
         otterbrix::cursor
         otterbrix::session
         otterbrix::document
-        otterbrix::dispatcher
         otterbrix::wal
         otterbrix::disk
         otterbrix::locks
