+++ integration/cpp/CMakeLists.txt
@@ -14,7 +14,6 @@ set(otterbrix_LIBS
         otterbrix::cursor
         otterbrix::session
         otterbrix::document
-        otterbrix::dispatcher
         otterbrix::wal
         otterbrix::disk
         otterbrix::locks
