+++ integration/cpp/base_spaces.cpp
@@ -4,13 +4,10 @@
 #include <core/system_command.hpp>
 #include <memory>
 #include <services/disk/manager_disk.hpp>
-#include <services/dispatcher/dispatcher.hpp>
 #include <services/memory_storage/memory_storage.hpp>
 #include <services/wal/manager_wal_replicate.hpp>
 
 namespace otterbrix {
-
-    using services::dispatcher::manager_dispatcher_t;
 
     constexpr static auto name_dispatcher = "dispatcher";
 
@@ -53,42 +50,29 @@ namespace otterbrix {
         memory_storage_ = actor_zeta::spawn_supervisor<services::memory_storage_t>(resource, scheduler_.get(), log_);
         trace(log_, "memory_storage finish");
 
-        trace(log_, "manager_dispatcher start");
-        manager_dispatcher_ =
-            actor_zeta::spawn_supervisor<services::dispatcher::manager_dispatcher_t>(resource,
-                                                                                     scheduler_dispather_.get(),
-                                                                                     log_);
-        trace(log_, "manager_dispatcher finish");
-
-        wrapper_dispatcher_ = std::make_unique<wrapper_dispatcher_t>(resource, manager_dispatcher_->address(), log_);
-        trace(log_, "manager_dispatcher create dispatcher");
-
-        actor_zeta::send(
-            manager_dispatcher_,
-            actor_zeta::address_t::empty_address(),
-            core::handler_id(core::route::sync),
-            std::make_tuple(memory_storage_->address(), manager_wal_->address(), manager_disk_->address()));
+        trace(log_, "wrapper_dispatcher start");
+        wrapper_dispatcher_ = std::make_unique<wrapper_dispatcher_t>(resource, memory_storage_->address(), log_);
+        trace(log_, "wrapper_dispatcher finish");
 
         actor_zeta::send(manager_wal_,
                          actor_zeta::address_t::empty_address(),
                          core::handler_id(core::route::sync),
-                         std::make_tuple(manager_disk_->address(), manager_dispatcher_->address()));
+                         std::make_tuple(manager_disk_->address(), memory_storage_->address()));
 
         actor_zeta::send(manager_disk_,
                          actor_zeta::address_t::empty_address(),
                          core::handler_id(core::route::sync),
-                         std::make_tuple(manager_dispatcher_->address()));
+                         std::make_tuple(memory_storage_->address()));
 
         actor_zeta::send(memory_storage_,
                          actor_zeta::address_t::empty_address(),
                          core::handler_id(core::route::sync),
-                         std::make_tuple(manager_dispatcher_->address(), manager_disk_->address()));
+                         std::make_tuple(wrapper_dispatcher_->address(), manager_disk_->address(), manager_wal_->address()));
 
         actor_zeta::send(manager_wal_, actor_zeta::address_t::empty_address(), wal::handler_id(wal::route::create));
         actor_zeta::send(manager_disk_,
                          actor_zeta::address_t::empty_address(),
                          disk::handler_id(disk::route::create_agent));
-        manager_dispatcher_->create_dispatcher(name_dispatcher);
         scheduler_dispather_->start();
         scheduler_->start();
         trace(log_, "spaces::spaces() final");
