+++ integration/cpp/route.hpp
@@ -2,16 +2,10 @@
 
 #include <services/collection/route.hpp>
 #include <services/disk/route.hpp>
-#include <services/dispatcher/route.hpp>
 #include <services/memory_storage/route.hpp>
 #include <services/wal/route.hpp>
 
 namespace otterbrix {
-
-    namespace dispatcher {
-        using services::dispatcher::handler_id;
-        using services::dispatcher::route;
-    } // namespace dispatcher
 
     namespace memory_storage {
         using services::memory_storage::handler_id;
