+++ integration/python/CMakeLists.txt
@@ -30,7 +30,6 @@ set(otterbrix_LIBS
         otterbrix::cursor
         otterbrix::session
         otterbrix::document
-        otterbrix::dispatcher
         otterbrix::wal
         otterbrix::disk
         otterbrix::locks
