+++ integration/python/spaces.hpp
@@ -6,7 +6,6 @@
 #include <components/log/log.hpp>
 
 #include <services/disk/manager_disk.hpp>
-#include <services/dispatcher/dispatcher.hpp>
 #include <services/memory_storage/memory_storage.hpp>
 #include <services/wal/manager_wal_replicate.hpp>
