+++ services/CMakeLists.txt
@@ -3,7 +3,6 @@ include_directories(${CMAKE_SOURCE_DIR})
 if(WASM_BUILD)
     add_subdirectory(wasm)
 else()
-    add_subdirectory(dispatcher)
     add_subdirectory(memory_storage)
     add_subdirectory(collection)
     add_subdirectory(wal)
