+++ services/collection/create_index_utils.cpp
@@ -78,7 +78,7 @@ namespace services::collection {
 
     void process_pending_indexes(context_collection_t* context) {
         // If doc exist and pending indexes not empty -> add curr ql to indexes and run loop
-        trace(context->log(), "Process pending indexes");
+        trace(context->log(), "Process pending indexes; size: {}", context->pending_indexes().size());
 
         assert(!(context->storage().empty()) && "Require non empty document storage");
 
@@ -93,6 +93,7 @@ namespace services::collection {
                 continue;
             }
             create_pending_index_impl(context, pipeline_context.get(), std::move(index));
+            trace(context->log(), "Process pending indexes; index proccessed");
         }
         context->storage().clear();
     }
