+++ services/memory_storage/memory_storage.cpp
@@ -9,6 +9,9 @@
 #include <core/tracy/tracy.hpp>
 #include <services/collection/collection.hpp>
 #include <services/collection/planner/create_plan.hpp>
+#include <components/translator/ql_translator.hpp>
+#include <services/disk/route.hpp>
+#include <services/wal/route.hpp>
 
 using namespace components::cursor;
 
@@ -16,6 +19,7 @@ namespace services {
 
     using namespace services::memory_storage;
     using namespace core::pmr;
+    using namespace components::ql;
 
     memory_storage_t::load_buffer_t::load_buffer_t(std::pmr::memory_resource* resource)
         : collections(resource) {}
@@ -24,6 +28,7 @@ namespace services {
                                        actor_zeta::scheduler_raw scheduler,
                                        log_t& log)
         : actor_zeta::cooperative_supervisor<memory_storage_t>(resource, "memory_storage")
+        , resource_(resource)
         , log_(log.clone())
         , e_(scheduler)
         , databases_(resource)
@@ -35,8 +40,7 @@ namespace services {
             resource,
             std::move(log_.clone()));
         add_handler(core::handler_id(core::route::sync), &memory_storage_t::sync);
-        add_handler(handler_id(route::execute_plan), &memory_storage_t::execute_plan);
-        add_handler(handler_id(route::load), &memory_storage_t::load);
+        add_handler(memory_storage::handler_id(memory_storage::route::execute_ql), &memory_storage_t::execute_ql);
 
         add_handler(collection::handler_id(collection::route::create_documents_finish),
                     &memory_storage_t::create_documents_finish_);
@@ -44,6 +48,11 @@ namespace services {
                     &memory_storage_t::execute_plan_finish_);
         add_handler(collection::handler_id(collection::route::size), &memory_storage_t::size);
         add_handler(collection::handler_id(collection::route::close_cursor), &memory_storage_t::close_cursor);
+
+        add_handler(core::handler_id(core::route::load), &memory_storage_t::load);
+        add_handler(disk::handler_id(disk::route::load_finish), &memory_storage_t::load_from_disk_result);
+        add_handler(wal::handler_id(wal::route::load_finish), &memory_storage_t::load_from_wal_result);
+        add_handler(wal::handler_id(wal::route::success), &memory_storage_t::wal_success);
     }
 
     memory_storage_t::~memory_storage_t() {
@@ -52,58 +61,264 @@ namespace services {
     }
 
     void memory_storage_t::sync(const address_pack& pack) {
-        manager_dispatcher_ = std::get<static_cast<uint64_t>(unpack_rules::manager_dispatcher)>(pack);
+        dispatcher_ = std::get<static_cast<uint64_t>(unpack_rules::dispatcher)>(pack);
         manager_disk_ = std::get<static_cast<uint64_t>(unpack_rules::manager_disk)>(pack);
+        manager_wal_ = std::get<static_cast<uint64_t>(unpack_rules::manager_wal)>(pack);
     }
 
-    void memory_storage_t::execute_plan(components::session::session_id_t& session,
-                                        components::logical_plan::node_ptr logical_plan,
-                                        components::ql::storage_parameters parameters) {
+    void memory_storage_t::execute_ql(components::session::session_id_t& session,
+                                        components::ql::ql_statement_t* ql,
+                                        actor_zeta::base::address_t address) {
+        sessions_.emplace(session, session_t{address, ql});
+        auto logic_plan = create_logic_plan(ql);
         using components::logical_plan::node_type;
 
-        switch (logical_plan->type()) {
+        switch (logic_plan.first->type()) {
             case node_type::create_database_t:
-                create_database_(session, std::move(logical_plan));
+                create_database_(session, std::move(logic_plan.first));
                 break;
             case node_type::drop_database_t:
-                drop_database_(session, std::move(logical_plan));
+                drop_database_(session, std::move(logic_plan.first));
                 break;
             case node_type::create_collection_t:
-                create_collection_(session, std::move(logical_plan));
+                create_collection_(session, std::move(logic_plan.first));
                 break;
             case node_type::drop_collection_t:
-                drop_collection_(session, std::move(logical_plan));
+                drop_collection_(session, std::move(logic_plan.first));
                 break;
             default:
-                execute_plan_(session, std::move(logical_plan), std::move(parameters));
+                execute_plan_(session, std::move(logic_plan.first), std::move(logic_plan.second));
                 break;
         }
     }
 
-    void memory_storage_t::size(components::session::session_id_t& session, collection_full_name_t&& name) {
-        trace(log_, "collection {}::{}::size", name.database, name.collection);
-        if (!check_collection_(session, name)) {
+    
+    void memory_storage_t::execute_ql_finish(components::session::session_id_t& session, cursor_t_ptr result) {
+        result_storage_[session] = result;
+        auto& s = sessions_.at(session);
+        auto* ql = s.get<ql_statement_t*>();
+        trace(log_,
+              "memory_storage_t::execute_ql_finish: session {}, {}, {}",
+              session.data(),
+              ql->to_string(),
+              result->is_success());
+        if (result->is_success()) {
+            //todo: delete
+
+            switch (ql->type()) {
+                case statement_type::create_database: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    actor_zeta::send(manager_disk_,
+                                     address(),
+                                     disk::handler_id(disk::route::append_database),
+                                     session,
+                                     ql->database_);
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::create_database),
+                                         session,
+                                         *static_cast<create_database_t*>(ql));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::create_collection: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    actor_zeta::send(manager_disk_,
+                                     address(),
+                                     disk::handler_id(disk::route::append_collection),
+                                     session,
+                                     ql->database_,
+                                     ql->collection_);
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::create_collection),
+                                         session,
+                                         *static_cast<create_collection_t*>(ql));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::insert_one: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::insert_one_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::insert_one),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::insert_many: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::insert_many_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::insert_many),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+                case statement_type::update_one: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::update_one_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::update_one),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::update_many: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::update_many_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::update_many),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::delete_one: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::delete_one_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::delete_one),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::delete_many: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        assert(s.is_type<ql_statement_t*>() && "Doesn't holds ql_statement_t*");
+                        auto statement = *static_cast<components::ql::delete_many_t*>(s.get<ql_statement_t*>());
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::delete_many),
+                                         session,
+                                         std::move(statement));
+                        return;
+                    }
+                    break;
+                }
+
+                case statement_type::drop_collection: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    collection_full_name_t name(ql->database_, ql->collection_);
+                    actor_zeta::send(manager_disk_,
+                                     address(),
+                                     disk::handler_id(disk::route::remove_collection),
+                                     session,
+                                     ql->database_,
+                                     ql->collection_);
+                    if (s.address().get() == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    } else {
+                        actor_zeta::send(manager_wal_,
+                                         address(),
+                                         wal::handler_id(wal::route::drop_collection),
+                                         session,
+                                         components::ql::drop_collection_t(ql->database_, ql->collection_));
+                        return;
+                    }
+                    break;
+                }
+                case statement_type::create_index:
+                case statement_type::drop_index: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: {}", to_string(ql->type()));
+                    const auto session_address = s.address().get();
+                    if (session_address == manager_wal_.get()) {
+                        wal_success(session, last_wal_id_);
+                    }
+                    // We initiate session from this dispatcher,
+                    // for now it means: dispatcher was called from manager_disk after initiate loading from disk.
+                    if (session_address == address().get()) {
+                        sessions_.erase(session);
+                        return;
+                    }
+                    break;
+                }
+                default: {
+                    trace(log_, "memory_storage_t::execute_ql_finish: non processed type");
+                }
+            }
+
+            //end: delete
+        }
+    }
+
+    void memory_storage_t::size(components::session::session_id_t& session,
+                                const std::string& database_name,
+                                const std::string& collection_name) {
+        collection_full_name_t name(database_name, collection_name);
+        trace(log_, "memory_storage_t::size: collection: {}.{}", database_name, collection_name);
+        if (!is_exists_collection_(name)) {
+            actor_zeta::send(current_message()->sender(),
+                             address(),
+                             collection::handler_id(collection::route::size_finish),
+                             session,
+                             0);
             return;
         }
         auto collection = collections_.at(name).get();
         if (collection->dropped()) {
             actor_zeta::send(current_message()->sender(),
                              address(),
-                             handler_id(collection::route::size_finish),
+                             collection::handler_id(collection::route::size_finish),
                              session,
-                             make_cursor(std::pmr::get_default_resource(), error_code_t::collection_dropped));
+                             0);
         } else {
-            auto* sub_cursor = new sub_cursor_t(collection->resource(), collection->name());
-            for (const auto& doc : collection->storage()) {
-                sub_cursor->append(components::document::document_view_t(doc.second));
-            }
-            auto cursor = make_cursor(collection->resource());
-            cursor->push(sub_cursor);
             actor_zeta::send(current_message()->sender(),
                              address(),
-                             handler_id(collection::route::size_finish),
+                             collection::handler_id(collection::route::size_finish),
                              session,
-                             std::move(cursor));
+                             collection->storage().size());
         }
     }
 
@@ -116,22 +331,32 @@ namespace services {
         }
     }
 
-    void memory_storage_t::load(components::session::session_id_t& session, const disk::result_load_t& result) {
-        trace(log_, "memory_storage_t:load");
+    void memory_storage_t::load(components::session::session_id_t& session) {
+        trace(log_, "memory_storage_t::load, session: {}", session.data());
+        load_session_ = session;
+        sessions_.emplace(session, session_t(current_message()->sender()));
+        actor_zeta::send(manager_disk_, address(), disk::handler_id(disk::route::load), session);
+    }
+
+    void memory_storage_t::load_from_disk_result(components::session::session_id_t& session,
+                                                 disk::result_load_t&& result) {
+        load_result_ = std::move(result);
+        trace(log_, "memory_storage_t:load_from_disk_result");
         load_buffer_ = std::make_unique<load_buffer_t>(resource());
         auto count_collections = std::accumulate(
-            (*result).begin(),
-            (*result).end(),
+            (*load_result_).begin(),
+            (*load_result_).end(),
             0ul,
             [](size_t sum, const disk::result_database_t& database) { return sum + database.collections.size(); });
         if (count_collections > 0) {
-            sessions_.emplace(session, session_t{nullptr, current_message()->sender(), count_collections});
+            trace(log_, "memory_storage_t:load_from_disk_result: documents: {}", count_collections);
+            sessions_.at(session).count_answers() = count_collections;
         }
-        for (const auto& database : (*result)) {
+        for (const auto& database : (*load_result_)) {
             debug(log_, "memory_storage_t:load:create_database: {}", database.name);
             databases_.insert(database.name);
             for (const auto& collection : database.collections) {
-                debug(log_, "memory_storage_t:load:create_collection: {}", collection.name);
+                debug(log_, "memory_storage_t:load_from_disk_result:create_collection: {}", collection.name);
                 collection_full_name_t name(database.name, collection.name);
                 auto context = new collection::context_collection_t(std::pmr::get_default_resource(),
                                                                     name,
@@ -139,7 +364,7 @@ namespace services {
                                                                     log_.clone());
                 collections_.emplace(name, context);
                 load_buffer_->collections.emplace_back(name);
-                debug(log_, "memory_storage_t:load:fill_documents: {}", collection.documents.size());
+                debug(log_, "memory_storage_t:load_from_disk_result:fill_documents: {}", collection.documents.size());
                 actor_zeta::send(executor_address_,
                                  address(),
                                  collection::handler_id(collection::route::create_documents),
@@ -149,8 +374,7 @@ namespace services {
             }
         }
         if (count_collections == 0) {
-            actor_zeta::send(current_message()->sender(), address(), handler_id(route::load_finish), session);
-            load_buffer_.reset();
+            load_from_disk_finished(session);
         }
     }
 
@@ -161,6 +385,181 @@ namespace services {
         std::unique_lock<spin_lock> _(lock_);
         set_current_message(std::move(msg));
         execute(this, current_message());
+    }
+
+    std::pair<components::logical_plan::node_ptr, components::ql::storage_parameters>
+    memory_storage_t::create_logic_plan(ql_statement_t* statement) {
+        //todo: cache logical plans
+        auto logic_plan = components::translator::ql_translator(resource_, statement);
+        auto parameters = statement->is_parameters()
+                              ? static_cast<components::ql::ql_param_statement_t*>(statement)->take_parameters()
+                              : components::ql::storage_parameters{};
+        return {logic_plan, parameters};
+    }
+
+    void memory_storage_t::wal_success(components::session::session_id_t& session, services::wal::id_t wal_id) {
+        trace(log_, "memory_storage_t::wal_success session : {}, wal id: {}", session.data(), wal_id);
+        actor_zeta::send(manager_disk_, address(), disk::handler_id(disk::route::flush), session, wal_id);
+
+        if (sessions_.contains(session)) {
+            return;
+        }
+
+        auto session_obj = sessions_.at(session);
+        const bool is_from_wal = session_obj.address().get() == manager_wal_.get();
+        if (is_from_wal) {
+            return;
+        }
+
+        trace(log_, "memory_storage_t::wal_success remove session : {}, wal id: {}", session.data(), wal_id);
+        auto result = result_storage_[session];
+        actor_zeta::send(session_obj.address(),
+                         address(),
+                         handler_id(memory_storage::route::execute_ql_finish),
+                         session,
+                         result);
+        sessions_.erase(session);
+        result_storage_.erase(session);
+    }
+    
+    void memory_storage_t::load_from_wal_result(components::session::session_id_t& session,
+                                            std::vector<services::wal::record_t>& in_records) {
+        // TODO think what to do with records
+        records_ = std::move(in_records);
+        load_count_answers_ = records_.size();
+        trace(log_,
+              "memory_storage_t::load_from_wal_result, session: {}, count commands: {}",
+              session.data(),
+              load_count_answers_);
+        if (load_count_answers_ == 0) {
+            trace(log_, "memory_storage_t::load_from_wal_result - empty records_");
+            actor_zeta::send(sessions_.at(session).address(),
+                             address(),
+                             core::handler_id(core::route::load_finish));
+            sessions_.erase(session);
+            return;
+        }
+        last_wal_id_ = records_[load_count_answers_ - 1].id;
+        for (auto& record : records_) {
+            switch (record.type) {
+                case statement_type::create_database: {
+                    assert(std::holds_alternative<create_database_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::create_database] variant "
+                           "record.data holds the alternative create_database_t");
+                    auto& data = std::get<create_database_t>(record.data);
+                    components::session::session_id_t session_database;
+                    execute_ql(session_database, &data, manager_wal_);
+                    break;
+                }
+                case statement_type::drop_database: {
+                    assert(std::holds_alternative<drop_database_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::drop_database] variant "
+                           "record.data holds the alternative drop_database_t");
+                    auto& data = std::get<drop_database_t>(record.data);
+                    components::session::session_id_t session_database;
+                    execute_ql(session_database, &data, manager_wal_);
+                    break;
+                }
+                case statement_type::create_collection: {
+                    assert(std::holds_alternative<create_collection_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::create_collection] variant "
+                           "record.data holds the alternative create_collection_t");
+                    auto& data = std::get<create_collection_t>(record.data);
+                    components::session::session_id_t session_collection;
+                    execute_ql(session_collection, &data, manager_wal_);
+                    break;
+                }
+                case statement_type::drop_collection: {
+                    assert(std::holds_alternative<drop_collection_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::drop_collection] variant "
+                           "record.data holds the alternative drop_collection_t");
+                    auto& data = std::get<drop_collection_t>(record.data);
+                    components::session::session_id_t session_collection;
+                    execute_ql(session_collection, &data, manager_wal_);
+                    break;
+                }
+                case statement_type::insert_one: {
+                    trace(log_, "memory_storage_t::load_from_wal_result: insert_one {}", session.data());
+                    assert(std::holds_alternative<insert_one_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::insert_one] variant "
+                           "record.data holds the alternative insert_one_t");
+                    auto& ql = std::get<insert_one_t>(record.data);
+                    components::session::session_id_t session_insert;
+                    execute_ql(session_insert, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::insert_many: {
+                    trace(log_, "memory_storage_t::load_from_wal_result: insert_many {}", session.data());
+                    assert(std::holds_alternative<insert_many_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::insert_many] variant "
+                           "record.data holds the alternative insert_many_t");
+                    auto& ql = std::get<insert_many_t>(record.data);
+                    components::session::session_id_t session_insert;
+                    execute_ql(session_insert, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::delete_one: {
+                    assert(std::holds_alternative<delete_one_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::delete_one] variant "
+                           "record.data holds the alternative delete_one_t");
+                    auto& ql = std::get<delete_one_t>(record.data);
+                    components::session::session_id_t session_delete;
+                    execute_ql(session_delete, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::delete_many: {
+                    assert(std::holds_alternative<delete_many_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::delete_many] variant "
+                           "record.data holds the alternative delete_many_t");
+                    auto& ql = std::get<delete_many_t>(record.data);
+                    components::session::session_id_t session_delete;
+                    execute_ql(session_delete, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::update_one: {
+                    trace(log_, "memory_storage_t::load_from_wal_result: update_one {}", session.data());
+                    assert(std::holds_alternative<update_one_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::update_one] variant "
+                           "record.data holds the alternative update_one_t");
+                    auto& ql = std::get<update_one_t>(record.data);
+                    components::session::session_id_t session_update;
+                    execute_ql(session_update, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::update_many: {
+                    assert(std::holds_alternative<update_many_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::update_many] variant "
+                           "record.data holds the alternative update_many_t");
+                    auto& ql = std::get<update_many_t>(record.data);
+                    components::session::session_id_t session_update;
+                    execute_ql(session_update, &ql, manager_wal_);
+                    break;
+                }
+                case statement_type::create_index: {
+                    assert(std::holds_alternative<create_index_t>(record.data) &&
+                           "[memory_storage_t::load_from_wal_result]: [ case: statement_type::create_index] variant "
+                           "record.data holds the alternative create_index_t");
+                    auto& ql = std::get<create_index_t>(record.data);
+                    components::session::session_id_t session_create_index;
+                    execute_ql(session_create_index, &ql, manager_wal_);
+                    break;
+                }
+                // TODO no drop index?
+                default:
+                    break;
+            }
+        }
+    }
+
+    void memory_storage_t::load_from_disk_finished(components::session::session_id_t& session) {
+        auto s = sessions_.at(session);
+        trace(log_, "memory_storage_t:load_from_disk_finished: {}, rest: {}", session.data(), s.count_answers());
+        actor_zeta::send(manager_disk_, address(), disk::handler_id(disk::route::load_indexes), session);
+        actor_zeta::send(manager_wal_, address(), wal::handler_id(wal::route::load), session, load_result_.wal_id());
+        load_result_.clear();
+        load_buffer_.reset();
+        actor_zeta::send(s.address(), address(), core::handler_id(core::route::load_finish), session);
+        sessions_.erase(session);
     }
 
     bool memory_storage_t::is_exists_database_(const database_name_t& name) const {
@@ -175,7 +574,7 @@ namespace services {
         if (!is_exists_database_(name)) {
             actor_zeta::send(current_message()->sender(),
                              this->address(),
-                             handler_id(route::execute_plan_finish),
+                             handler_id(memory_storage::route::execute_ql_finish),
                              session,
                              make_cursor(default_resource(), error_code_t::database_not_exists, "database not exists"));
             return false;
@@ -190,7 +589,7 @@ namespace services {
                 actor_zeta::send(
                     current_message()->sender(),
                     this->address(),
-                    handler_id(route::execute_plan_finish),
+                    handler_id(memory_storage::route::execute_ql_finish),
                     session,
                     make_cursor(default_resource(), error_code_t::collection_not_exists, "collection not exists"));
                 return false;
@@ -207,7 +606,7 @@ namespace services {
             actor_zeta::send(
                 current_message()->sender(),
                 this->address(),
-                handler_id(route::execute_plan_finish),
+                handler_id(memory_storage::route::execute_ql_finish),
                 session,
                 make_cursor(default_resource(), error_code_t::database_already_exists, "database already exists"));
             return;
@@ -215,7 +614,7 @@ namespace services {
         databases_.insert(logical_plan->database_name());
         actor_zeta::send(current_message()->sender(),
                          this->address(),
-                         handler_id(route::execute_plan_finish),
+                         handler_id(memory_storage::route::execute_ql_finish),
                          session,
                          make_cursor(default_resource(), operation_status_t::success));
     }
@@ -227,7 +626,7 @@ namespace services {
             databases_.erase(logical_plan->database_name());
             actor_zeta::send(current_message()->sender(),
                              this->address(),
-                             handler_id(route::execute_plan_finish),
+                             handler_id(memory_storage::route::execute_ql_finish),
                              session,
                              make_cursor(default_resource(), operation_status_t::success));
         }
@@ -240,7 +639,7 @@ namespace services {
             if (is_exists_collection_(logical_plan->collection_full_name())) {
                 actor_zeta::send(current_message()->sender(),
                                  this->address(),
-                                 handler_id(route::execute_plan_finish),
+                                 handler_id(memory_storage::route::execute_ql_finish),
                                  session,
                                  make_cursor(default_resource(),
                                              error_code_t::collection_already_exists,
@@ -255,7 +654,7 @@ namespace services {
             auto cursor = make_cursor(default_resource(), operation_status_t::success);
             actor_zeta::send(current_message()->sender(),
                              this->address(),
-                             handler_id(route::execute_plan_finish),
+                             handler_id(memory_storage::route::execute_ql_finish),
                              session,
                              cursor);
         }
@@ -265,16 +664,14 @@ namespace services {
                                             components::logical_plan::node_ptr logical_plan) {
         trace(log_, "memory_storage_t:drop_collection {}", logical_plan->collection_full_name().to_string());
         if (check_collection_(session, logical_plan->collection_full_name())) {
-            sessions_.emplace(session, session_t{logical_plan, current_message()->sender(), 1});
             actor_zeta::send(
                 current_message()->sender(),
                 address(),
-                handler_id(route::execute_plan_finish),
+                handler_id(memory_storage::route::execute_ql_finish),
                 session,
                 collections_.at(logical_plan->collection_full_name())->drop()
                     ? make_cursor(default_resource(), operation_status_t::success)
                     : make_cursor(default_resource(), error_code_t::other_error, "collection not dropped"));
-            sessions_.erase(session);
             collections_.erase(logical_plan->collection_full_name());
             trace(log_, "memory_storage_t:drop_collection_finish {}", logical_plan->collection_full_name().to_string());
         }
@@ -293,15 +690,10 @@ namespace services {
             collection_full_name_t name =
                 dependency_tree_collections_names.extract(dependency_tree_collections_names.begin()).value();
             if (!check_collection_(session, name)) {
-                trace(log_,
-                      "memory_storage_t:execute_plan_: collection not found {}, sesion: {}",
-                      name.to_string(),
-                      session.data());
                 return;
             }
             collections_context_storage.emplace(std::move(name), collections_.at(name).get());
         }
-        sessions_.emplace(session, session_t{logical_plan, current_message()->sender(), 1});
         actor_zeta::send(executor_address_,
                          address(),
                          collection::handler_id(collection::route::execute_plan),
@@ -313,11 +705,11 @@ namespace services {
 
     void memory_storage_t::execute_plan_finish_(components::session::session_id_t& session, cursor_t_ptr result) {
         auto& s = sessions_.at(session);
-        debug(log_,
+        trace(log_,
               "memory_storage_t:execute_plan_finish: session: {}, success: {}",
               session.data(),
               result->is_success());
-        actor_zeta::send(s.sender, address(), handler_id(route::execute_plan_finish), session, std::move(result));
+        actor_zeta::send(s.address(), address(), memory_storage::handler_id(memory_storage::route::execute_ql_finish), session, std::move(result));
         sessions_.erase(session);
     }
 
@@ -326,12 +718,10 @@ namespace services {
             return;
         }
         auto& s = sessions_.at(session);
-        --s.count_answers;
-        debug(log_, "memory_storage_t:create_documents_finish: {}, rest: {}", session.data(), s.count_answers);
-        if (s.count_answers == 0) {
-            actor_zeta::send(s.sender, address(), handler_id(route::load_finish), session);
-            load_buffer_.reset();
-            sessions_.erase(session);
+        --s.count_answers();
+        trace(log_, "memory_storage_t:create_documents_finish: {}, rest: {}", session.data(), s.count_answers());
+        if (s.count_answers() == 0) {
+            load_from_disk_finished(session);
         }
     }
